requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-firestore
firebase-admin
python-dotenv
twilio
openai
protobuf